Adds User model, controller, and specs
Lots done here. Most of it scaffolding. Added a User model, views and
the beginning of specs. There is something wonky with the application
views currently...only the generic rails welcome page is displayed. That
will be addressed in a later commit. For now, bear with the partially
functional app and failing tests.